gemini_browser_ui/requirements.txt
# Gemini Browser UI - Requirements
# Standalone micro-project for Gemini Computer Use with embedded browser

# Core AI/ML
google-genai>=1.0.0

# Browser automation
playwright>=1.40.0

# Web framework
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# WebSocket support
python-socketio>=5.10.0
eventlet>=0.33.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0

# OAuth 2.0 / Authentication
authlib>=1.3.0
flask-session>=0.6.0

# Optional: For development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
